moved the header down into the content of the homescreen for gestures to make more sense
